Add DropDownController component for SingleDatePicker
